+++ src/main/kotlin/dev/protsenko/securityLinter/kubernetes/quickfix/ReplacedValueEnum.kt
@@ -1,0 +1,5 @@
+package dev.protsenko.securityLinter.kubernetes.quickfix
+
+enum class ReplacedValueEnum(val value: String) {
+    TRUE("true"), FALSE("false"), ONE_THOUSAND("1000")
+}